+++ Example/Emission/ARRootViewController.m
@@ -375,7 +375,7 @@
 - (ARCellData *)jumpToFair
 {
   return [self tappableCellDataWithTitle:@"Fair" selection:^{
-    id viewController = [[ARFairComponentViewController alloc] initWithFairID:@"art-los-angeles-contemporary"];
+    id viewController = [[ARFairComponentViewController alloc] initWithFairID:@"frieze-los-angeles-2019"];
     [self.navigationController pushViewController:viewController animated:YES];
   }];
 }